Moved All Movement Input to Player
